added new callback functions to be able to used by aws
